+++ frontend/src/metabase/metabot/components/AIMarkdown/AIMarkdown.tsx
@@ -11,8 +11,8 @@ import {
 } from "react";
 import { t } from "ttag";
 
-import { parseMetabaseProtocolLink } from "metabase/metabot/utils/links";
 import { ActionIcon, CopyButton, Icon, Tooltip } from "metabase/ui";
+import { parseMetabaseProtocolLink } from "metabase/urls";
 
 import S from "./AIMarkdown.module.css";
 import { StreamingMarkdown } from "./StreamingMarkdown";
