+++ frontend/src/metabase/metabot/components/AIMarkdown/components/MarkdownSmartLink.tsx
@@ -4,14 +4,15 @@ import { match } from "ts-pattern";
 import { EntityIcon } from "metabase/common/components/EntityIcon";
 import { useGetIcon } from "metabase/hooks/use-icon";
 import { getConversationChart } from "metabase/metabot/state";
-import type {
-  MetabaseProtocolEntityModel,
-  ParsedMetabaseProtocolLink,
-} from "metabase/metabot/utils/links";
 import { useSelector } from "metabase/redux";
 import { useEntityData } from "metabase/rich_text_editing/tiptap/extensions/SmartLink/SmartLinkNode";
 import { entityToUrlableModel } from "metabase/rich_text_editing/tiptap/extensions/shared/suggestionUtils";
-import { conversationChartUrl, modelToUrl } from "metabase/urls";
+import {
+  type MetabaseProtocolEntityModel,
+  type ParsedMetabaseProtocolLink,
+  conversationChartUrl,
+  modelToUrl,
+} from "metabase/urls";
 
 import S from "../AIMarkdown.module.css";
