+++ frontend/src/metabase/urls/index.ts
@@ -11,6 +11,7 @@ export * from "./data-studio";
 export * from "./dependencies";
 export * from "./documents";
 export * from "./indexed-entities";
+export * from "./metabase-protocol";
 export * from "./metabot";
 export * from "./metrics";
 export * from "./misc";
